Update main file to use src/ path.
Removed the dist path because there will be no uglification or modifying of
code for release. As this is a Node module not meant for the browser we
do not need to use a separate dist/ directory for the built components.
We do use a src/ dir just to keep the configuration stuff separate from
our actual project code.